+++ inst/app/modules/nvc_analysis/report_server.R
@@ -27,6 +27,7 @@ report <- function(input, output, session,
     
     filename = function() {
       paste0("RMAVIS.Report.",
+             "v1-1-0.",
              format(Sys.time(), "%y-%m-%d.%H-%M-%S"),
              ".pdf",
              sep="")