+++ inst/app/modules/ui.R
@@ -14,8 +14,6 @@ ui <- bslib::page_navbar(
       # App Title
       shiny::div(shiny::h1("RMAVIS")),
       
-      # bslib::card_image(file = "www/nationalReference.png", fill = FALSE, max_height = "62px"),
-      
       # Align cell contents in the middle vertically
       cellArgs = list(style = "vertical-align: middle !important; color: #565656 !important;")